@@ -3,19 +3,20 @@ description: Sound meter
 
 publish_to: "none"
 
-version: 1.0.1+2
+version: 1.1.0-f+3
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # cupertino_icons: ^1.0.0
-  noise_meter: ^2.0.0
-  syncfusion_flutter_charts: ^18.3.50
-  url_launcher: ^5.7.10
+  noise_meter: ^3.0.1
+  syncfusion_flutter_charts: ^19.1.67
+  url_launcher: ^6.0.6
+  google_fonts: ^2.1.0
 
 dev_dependencies:
   flutter_test:
